All File
How to Develop
Step 1 : Create New Project
Step 2 Define a layout file
Step 3 Create an XML file (WidgetDemo) which describes the properties of
the widget, e.g. size or the fixed update frequency.
Step 4 :Create a BroadcastReceiver which is used to build the user
interface of the Widget.
Step 5 : Enter the Widget configuration in the AndroidManifest.xml file.